# CI: improve composer cache performance