--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -39,6 +39,7 @@ jobs:
   # @todo: common initialization: 'composer install' for the site-under-test
   build:
     <<: *defaults
+    resource_class: large
     steps:
       - checkout