--- a/.circleci/scripts/pantheon/02-init-site-under-test-clone-existing
+++ b/.circleci/scripts/pantheon/02-init-site-under-test-clone-existing
@@ -4,6 +4,7 @@ set -eo pipefail
 # delete (transient) npm dependencies, as they're only used to build assets here
 # and not needed for the Pantheon runtime
 rm -rf \
+  node_modules \
   web/modules/custom/*/node_modules \
   web/themes/custom/*/node_modules