--- a/config/translations/vaccine-sites.pot
+++ b/config/translations/vaccine-sites.pot
@@ -1,0 +1,43 @@
+#. English strings for /vaccine-sites
+#, c-format
+msgid ""
+msgstr ""
+"Content-Type: text/plain; charset=utf-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: nplurals=2; plural=(n > 1);\n"
+
+msgid "This pharmacy can only vaccinate kids 3 and older."
+msgstr ""
+
+msgid "Vaccines for kids"
+msgstr ""
+
+msgid "Age range"
+msgstr ""
+
+msgid "6 months to 4 years"
+msgstr ""
+
+msgid "5 to 11 years"
+msgstr ""
+
+msgid "12 to 17 years"
+msgstr ""
+
+msgid "5 and under"
+msgstr ""
+
+msgid "4 and under"
+msgstr ""
+
+msgid "Moderna"
+msgstr ""
+
+msgid "Pfizer"
+msgstr ""
+
+msgid "<span>from</span>"
+msgstr ""
+
+msgid "Apply"
+msgstr ""