--- a/dangerfile.ts
+++ b/dangerfile.ts
@@ -26,15 +26,16 @@ function checkTranslations (cwd: string) {
   for (const poFile of poFiles) {
     const base = basename(poFile)
     const parts = base.split('.')
-    const [lang] = parts.length >= 2 ? parts.slice(-2) : []
+    const [lang] = parts.length > 2 ? parts.slice(-2) : []
     if (!lang) {
       fail(`Missing language code in filename: \`${join(cwd, poFile)}\` should be: \`${join(cwd, base)}.(${translationLangs.join('|')}).po\`)`)
     } else if (!translationLangs.includes(lang)) {
       fail(`Invalid language code in ${poFile}: \`${lang}\`; expected one of \`${translationLangs.join('`, `')}\``)
-    }
-    const potFile = poFile.replace(`.${lang}.po`, '.pot')
-    if (!potFiles.includes(potFile) && !missingTemplates.includes(potFile)) {
-      missingTemplates.push(potFile)
+    } else {
+      const potFile = poFile.replace(lang ? `.${lang}.po` : '.po', '.pot')
+      if (!potFiles.includes(potFile) && !missingTemplates.includes(potFile)) {
+        missingTemplates.push(potFile)
+      }
     }
   }
 
@@ -43,11 +44,11 @@ function checkTranslations (cwd: string) {
   }
 
   for (const potFile of potFiles) {
-    for (const lang of translationLangs) {
-      const poFile = potFile.replace(/\.pot$/, `.${lang}.po`)
-      if (!poFiles.includes(poFile)) {
-        fail(`Missing ${locales[lang]} translation for \`${potFile}\`: expected \`${poFile}\``)
-      }
+    const missingTranslations = translationLangs
+      .map(lang => ({ lang, path: potFile.replace(/\.pot$/, `.${lang}.po`) }))
+      .filter(({ path }) => !poFiles.includes(path))
+    if (missingTranslations.length > 0) {
+      fail(`Missing translations for \`${potFile}\`: ${missingTranslations.map(({ lang, path }) => `${lang} (\`${path}\`)`)}`)
     }
   }
 }