--- a/dangerfile.ts
+++ b/dangerfile.ts
@@ -28,9 +28,9 @@ function checkTranslations (cwd: string) {
     const parts = base.split('.')
     const [lang] = parts.length > 2 ? parts.slice(-2) : []
     if (!lang) {
-      fail(`Missing language code in filename: \`${join(cwd, poFile)}\` should be: \`${join(cwd, base)}.(${translationLangs.join('|')}).po\`)`)
+      fail(`Missing language code in filename: ${code(join(cwd, poFile))} should be named ${code(`${base}.{${translationLangs.join(',')}}.po`)}`)
     } else if (!translationLangs.includes(lang)) {
-      fail(`Invalid language code in ${poFile}: \`${lang}\`; expected one of \`${translationLangs.join('`, `')}\``)
+      fail(`Invalid language code "${lang}" in ${code(join(cwd, poFile))} (expected one of ${glue(translationLangs.map(code), ', ', ', or ')}`)
     } else {
       const potFile = poFile.replace(lang ? `.${lang}.po` : '.po', '.pot')
       if (!potFiles.includes(potFile) && !missingTemplates.includes(potFile)) {
@@ -40,7 +40,7 @@ function checkTranslations (cwd: string) {
   }
 
   for (const potFile of missingTemplates) {
-    fail(`Missing English template \`${potFile}\` for \`${potFile.replace('.pot', '.*.po')}\``)
+    fail(`Missing English template ${code(potFile)} for ${code(potFile.replace('.pot', '.*.po'))}`)
   }
 
   for (const potFile of potFiles) {
@@ -48,7 +48,21 @@ function checkTranslations (cwd: string) {
       .map(lang => ({ lang, path: potFile.replace(/\.pot$/, `.${lang}.po`) }))
       .filter(({ path }) => !poFiles.includes(path))
     if (missingTranslations.length > 0) {
-      fail(`Missing translations for \`${potFile}\`: ${missingTranslations.map(({ lang, path }) => `${lang} (\`${path}\`)`)}`)
+      fail(`Missing translations for ${code(potFile)}: ${glue(missingTranslations.map(({ lang, path }) => `${code(path)} (${locales[lang]}))`), ', ', ', and ')}`)
     }
   }
+}
+
+function glue (parts: string[], glue, lastGlue = glue) {
+  const { length } = parts
+  if (length < 2) return parts.join(glue)
+  const last = length - 1
+  const secondToLast = last - 1
+  return parts.flatMap((part, i) => {
+    return i === last ? part : [part, i === secondToLast ? lastGlue : glue]
+  }).join('')
+}
+
+function code (str: string) {
+  return `\`${str}\``
 }