… .github/workflows/checks.yml
… .github/workflows/danger.yml
… composer-manifest.yaml
… composer.lock
… config/core.entity_view_display.media.image.captioned.yml
… config/core.entity_view_display.media.image.default.yml
… config/core.entity_view_display.media.image.event.yml
… config/core.entity_view_display.media.image.news.yml
… config/core.entity_view_display.media.image.preview.yml
… config/core.entity_view_display.node.campaign.default.yml
… config/core.entity_view_display.node.form_confirmation_page.default.yml
… config/core.entity_view_display.node.person.default.yml
… config/core.entity_view_display.user.user.compact.yml
… config/core.entity_view_display.user.user.default.yml
… config/core.extension.yml
… config/language/es/views.view.archive.yml
… config/language/es/views.view.block_content.yml
… config/language/es/views.view.content.yml
… config/language/es/views.view.content_recent.yml
… config/language/es/views.view.fancy_file_delete.yml
… config/language/es/views.view.fancy_file_list_unmanaged.yml
… config/language/es/views.view.files.yml
… config/language/es/views.view.frontpage.yml
… config/language/es/views.view.glossary.yml
… config/language/es/views.view.locked_content.yml
… config/language/es/views.view.media.yml
… config/language/es/views.view.moderated_content.yml
… config/language/es/views.view.redirect.yml
… config/language/es/views.view.scheduler_scheduled_content.yml
… config/language/es/views.view.taxonomy_term.yml