I noticed in [this build](https://app.circleci.com/pipelines/github/SFDigitalServices/sfgov/5004/workflows/71d9e0fc-a630-437c-abd7-d1930d285525/jobs/15944?invite=true#step-102-2) that Circle CI doesn't appear to be restoring the Composer cache:

```
Found a cache from build 15776 at composer-0Dn0ScT67box2O3MwW+PpgfIb+l4xAjEhIg13060uoE=
Size: 32 B
Cached paths:
  * /home/tester/example_drops_8_composer/$HOME/.composer/cache
```

I expected the `vendor` directory (which, AFAIK, is where Composer installs depdendencies) to be listed under `Cached paths`, thrashed a bit, and got it working (see [this `build` job](https://app.circleci.com/pipelines/github/SFDigitalServices/sfgov/5015/workflows/5b97a135-81ac-46cd-954c-f3c0ad131d0b/jobs/15970) and [this `build_deploy_and_test` job](https://app.circleci.com/pipelines/github/SFDigitalServices/sfgov/5015/workflows/5b97a135-81ac-46cd-954c-f3c0ad131d0b/jobs/15971?invite=true#step-103-5)). You can see in these jobs that the `build assets (composer)` step timing improves dramatically:

| Time | Commit | Cached? |
| :--- | :--- | :---
| [52s](https://app.circleci.com/pipelines/github/SFDigitalServices/sfgov/5008/workflows/91c22af6-7534-4e9b-a41b-8fe035d12519/jobs/15954?invite=true#step-105-1) | [484f941](https://github.com/SFDigitalServices/sfgov/commit/484f941ce3c3e572fbad90d821493439c4fc6d54) | no
| [49s](https://app.circleci.com/pipelines/github/SFDigitalServices/sfgov/5009/workflows/ef610399-11a7-4649-ae74-0929cf5740b2/jobs/15956?invite=true#step-105-1) | [f92fef9](https://github.com/SFDigitalServices/sfgov/commit/f92fef9b662c13a5756a390ba382e6166b227df5) | no
| [21s](https://app.circleci.com/pipelines/github/SFDigitalServices/sfgov/5009/workflows/eb1f1589-331a-4e85-8f50-fbe03d56bd41/jobs/15958?invite=true#step-106-1) | [f92fef9](https://github.com/SFDigitalServices/sfgov/commit/f92fef9b662c13a5756a390ba382e6166b227df5) | yes
| [17s](https://app.circleci.com/pipelines/github/SFDigitalServices/sfgov/5012/workflows/20bca97f-43a6-41ee-a9f6-d36debdbc4d3/jobs/15964?invite=true#step-106-1) | [a4202f8](https://github.com/SFDigitalServices/sfgov/commit/a4202f818e2d37c84034185084cd85bdce551626) | yes
| [18s](https://app.circleci.com/pipelines/github/SFDigitalServices/sfgov/5015/workflows/5b97a135-81ac-46cd-954c-f3c0ad131d0b/jobs/15971?invite=true#step-106-1) | [a4d3b86](https://github.com/SFDigitalServices/sfgov/commit/a4d3b860b07fc366bf3a18e71cd20a05b140fd4c) | yes

As of [a4d3b86](https://github.com/SFDigitalServices/sfgov/pull/1352/commits/a4d3b860b07fc366bf3a18e71cd20a05b140fd4c) the deployment to Pantheon is still failing, which I hope is unrelated. 😬